+++ src/ai_marketplace_monitor/cli.py
@@ -11,7 +11,6 @@ import typer
 from rich.logging import RichHandler
 
 from . import __version__
-from .monitor import MarketplaceMonitor
 from .utils import CacheType, amm_home, cache, counter, hilight
 
 app = typer.Typer()
@@ -130,6 +129,9 @@ def main(
         logger.info(f"""{hilight("[Clear Cache]", "succ")} Cache cleared.""")
         sys.exit(0)
 
+    # make --version a bit faster by lazy loading of MarketplaceMonitor
+    from .monitor import MarketplaceMonitor
+
     if items is not None:
         try:
             monitor = MarketplaceMonitor(config_files, headless, disable_javascript, logger)
@@ -150,6 +152,7 @@ def main(
         sys.exit(0)
     except Exception as e:
         logger.error(f"""{hilight("[Monitor]", "fail")} {e}""")
+        raise
         sys.exit(1)
     finally:
         monitor.stop_monitor()