+++ src/ai_marketplace_monitor/user.py
@@ -6,34 +6,28 @@ from typing import Any, List, Tuple, Type
 from diskcache import Cache  # type: ignore
 
 from .ai import AIResponse  # type: ignore
+from .email_notify import EmailNotificationConfig
 from .listing import Listing
 from .marketplace import TItemConfig
-from .pushbullet import PushbulletConfig
-from .smtp import SMTPConfig
-from .utils import (
-    CacheType,
-    CounterItem,
-    NotificationStatus,
-    cache,
-    convert_to_seconds,
-    counter,
-    hilight,
-)
+from .notification import NotificationConfig, NotificationStatus
+from .pushbullet import PushbulletNotificationConfig
+from .utils import CacheType, CounterItem, cache, convert_to_seconds, counter, hilight
 
 
 @dataclass
-class UserConfig(SMTPConfig, PushbulletConfig):
-    # this argument is required
-    email: List[str] | None = None
-    smtp: str | None = None
-    remind: int | None = None
+class UserConfig(EmailNotificationConfig, PushbulletNotificationConfig):
+    """UserConfiguration
 
-    def handle_pushbullet_token(self: "UserConfig") -> None:
-        if self.pushbullet_token is None:
-            return
-        if not isinstance(self.pushbullet_token, str) or not self.pushbullet_token:
-            raise ValueError("user requires an non-empty pushbullet_token.")
-        self.pushbullet_token = self.pushbullet_token.strip()
+    Derive from EmailNotificationConfig, PushbulletNotificationConfig allows
+    the user config class to use settings from both classes.
+
+    It is possible to dynamically added these classes as parent class
+    of UserConfig, but it is troublesome to make sure that these classes
+    are imported.
+    """
+
+    notify_with: List[str] | None = None
+    remind: int | None = None
 
     def handle_remind(self: "UserConfig") -> None:
         if self.remind is None:
@@ -65,24 +59,18 @@ class UserConfig(SMTPConfig, PushbulletConfig):
                 f"Item {hilight(self.name)} remind must be an time (e.g. 1 day) or false."
             )
 
-    def handle_email(self: "UserConfig") -> None:
-        if self.email is None:
+    def handle_notify_with(self: "UserConfig") -> None:
+        if self.notify_with is None:
             return
-        if isinstance(self.email, str):
-            self.email = [self.email]
-        if not isinstance(self.email, list) or not all(
-            (isinstance(x, str) and "@" in x and "." in x.split("@")[1]) for x in self.email
+
+        if isinstance(self.notify_with, str):
+            self.notify_with = [self.notify_with]
+
+        if not isinstance(self.notify_with, list) or not all(
+            isinstance(x, str) for x in self.notify_with
         ):
             raise ValueError(
-                f"Item {hilight(self.name)} email must be a string or list of string."
-            )
-
-    def handle_smtp(self: "UserConfig") -> None:
-        if self.smtp is None:
-            return
-        if not isinstance(self.smtp, str):
-            raise ValueError(
-                f"Item {hilight(self.name)} smtp must be a valid smtp server configuration name."
+                f"Item {hilight(self.name)} notify_with must be a list of notification section values."
             )
 
 
@@ -164,10 +152,9 @@ class User:
                 )
             return
         statuses = [self.notification_status(listing, local_cache) for listing in listings]
-        if self.config.notify_through_pushbullet(
+
+        if NotificationConfig.notify_all(
             listings, ratings, statuses, force=force, logger=self.logger
-        ) or self.config.notify_through_email(
-            self.config.email, listings, ratings, statuses, force=force, logger=self.logger
         ):
             counter.increment(CounterItem.NOTIFICATIONS_SENT, item_config.name)
             for listing, ns in zip(listings, statuses):