+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.0] - 2025-03-02
+
+- Reorganize notification settings
+- Support the use of environment variables for passwords
+- Support browser proxy
+
+**BREAKING CHANGES**
+
+- Rename `smtp` sections to `notification`
+- Rename parameter `smtp` to `notify_with`
+
 ## [0.7.11] - 2025-03-01
 
 - Fix a bug on the handling of logical expressions for `keywords` and `antikeywords`.